chore(asm): refactor functions to taint_tracking
