+++ ddtrace/appsec/_asm_request_context.py
@@ -54,58 +54,12 @@ def get_ip():  # type: () -> Optional[str]
     return _DD_EARLY_IP_CONTEXTVAR.get()
 
 
-def add_taint_pyobject(pyobject, op1, op2):
-    from ddtrace.appsec.iast._taint_tracking import new_pyobject_id
-
-    if not (is_pyobject_tainted(op1) or is_pyobject_tainted(op2)):
-        return pyobject
-
-    pyobject = new_pyobject_id(pyobject, len(pyobject))
-    taint_dict = _DD_IAST_TAINT_DICT.get()
-    new_ranges = []
-    if is_pyobject_tainted(op1):
-        new_ranges = list(taint_dict[id(op1)])
-    if is_pyobject_tainted(op2):
-        offset = len(op1)
-        for input_info, start, size in taint_dict[id(op2)]:
-            new_ranges.append((input_info, start + offset, size))
-
-    taint_dict[id(pyobject)] = tuple(new_ranges)
-    _DD_IAST_TAINT_DICT.set(taint_dict)
-    return pyobject
-
-
-def taint_pyobject(pyobject, input_info):
-    from ddtrace.appsec.iast._taint_tracking import new_pyobject_id
-
-    if not pyobject:  # len(pyobject) < 1
-        return pyobject
-    assert input_info is not None
-    len_pyobject = len(pyobject)
-    pyobject = new_pyobject_id(pyobject, len_pyobject)
-    taint_dict = _DD_IAST_TAINT_DICT.get()
-    taint_dict[id(pyobject)] = ((input_info, 0, len_pyobject),)
-    _DD_IAST_TAINT_DICT.set(taint_dict)
-    return pyobject
-
-
-def is_pyobject_tainted(pyobject):
-    return id(pyobject) in _DD_IAST_TAINT_DICT.get()
-
-
-def set_tainted_ranges(pyobject, ranges):
-    assert pyobject not in _DD_IAST_TAINT_DICT.get()
-    taint_dict = _DD_IAST_TAINT_DICT.get()
-    taint_dict[id(pyobject)] = ranges
+def set_taint_dict(taint_dict):  # type: (dict) -> None
     _DD_IAST_TAINT_DICT.set(taint_dict)
 
 
-def get_tainted_ranges(pyobject):
-    return _DD_IAST_TAINT_DICT.get().get(id(pyobject), tuple())
-
-
-def clear_taint_mapping():
-    _DD_IAST_TAINT_DICT.set({})
+def get_taint_dict():  # type: () -> dict
+    return _DD_IAST_TAINT_DICT.get()
 
 
 # Note: get/set headers use Any since we just carry the headers here without changing or using them
